Move midpoint into _utils.operations
